use basedir stuff stored on root node instead of dealing with it in this
2003-05-02  Havoc Pennington  <hp@redhat.com>

	* src/menu-process.c: use basedir stuff stored on root node
	instead of dealing with it in this file

	* src/menu-parser.c (menu_load): set basedir

	* src/menu-layout.c (menu_node_copy_one): copy fields in MenuNode
	"subclasses"
	(menu_node_get_basedir): new
	(menu_node_get_content_as_path): new

	* configure.in: add more compiler warnings, and --enable-tests

	* src/menu-layout.c (dfu_test_menu_nodes): start setting up unit
	test stuff

+++ src/desktop_file.c
@@ -1518,13 +1518,13 @@ static const char *
 get_canonical_encoding (const char *encoding)
 {
   int i;
-  for (i = 0; i < G_N_ELEMENTS (enc_aliases); i++)
+  for (i = 0; i < (int) G_N_ELEMENTS (enc_aliases); i++)
     {
       if (aliases_equal (enc_aliases[i].alias, encoding))
 	return enc_aliases[i].value;
     }
 
-  for (i = 0; i < G_N_ELEMENTS (known_encodings); i++)
+  for (i = 0; i < (int) G_N_ELEMENTS (known_encodings); i++)
     {
       if (aliases_equal (known_encodings[i].encoding, encoding))
 	return known_encodings[i].encoding;
@@ -1555,7 +1555,7 @@ get_encoding_from_lang (const char *lang)
 {
   int i, j;
   
-  for (i = 0; i < G_N_ELEMENTS (known_encodings); i++)
+  for (i = 0; i < (int) G_N_ELEMENTS (known_encodings); i++)
     {
       for (j = 0; j < N_LANG; j++)
 	{
